+++ src/bin/daemon/daemon.hpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2018  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2020  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -27,7 +27,6 @@
 #include <cstdio>
 #include <sys/types.h>
 
-#include <fstream>
 #include <string>
 
 #include "daemonControl.hpp"
@@ -51,8 +50,8 @@ class GLogiKDaemon
 		pid_t _pid = 0;
 		FILE* _LOGfd = nullptr;
 		std::string _pidFileName;
-		std::ofstream _pidFile;
 		NSGKDBus::GKDBus* _pDBus;
+		bool _PIDFileCreated;
 
 		void createPIDFile(void);
 		void dropPrivileges(void);