+++ src/bin/daemon/devicesManager.cpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2019  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2020  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -807,8 +807,9 @@ void DevicesManager::startMonitoring(NSGKDBus::GKDBus* pDBus) {
 
 			unsigned short c = 0;
 
-			while( DaemonControl::isDaemonRunning() ) {
-			int ret = poll(fds, 1, 100);
+			while( DaemonControl::isDaemonRunning() )
+			{
+				int ret = poll(fds, 1, 100);
 
 				// receive data ?
 				if( ret > 0 ) {