+++ src/bin/service/service.cpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2019  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2020  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -19,13 +19,12 @@
  *
  */
 
-#include <errno.h>
 #include <poll.h>
-#include <sys/types.h>
-#include <unistd.h>
 
 #include <cstring>
 #include <cstdlib>
+
+#include <syslog.h>
 
 #include <new>
 #include <fstream>
@@ -56,6 +55,8 @@ DesktopService::DesktopService() :
 	_verbose(false),
 	_pGKfs(nullptr)
 {
+	openlog(GLOGIKS_DESKTOP_SERVICE_NAME, LOG_PID|LOG_CONS, LOG_USER);
+
 	LOG_TO_FILE_AND_CONSOLE::ConsoleReportingLevel() = INFO;
 	if( LOG_TO_FILE_AND_CONSOLE::ConsoleReportingLevel() != NONE ) {
 		LOG_TO_FILE_AND_CONSOLE::ConsoleStream() = stderr;
@@ -63,40 +64,7 @@ DesktopService::DesktopService() :
 
 #if DEBUGGING_ON
 	LOG_TO_FILE_AND_CONSOLE::FileReportingLevel() = DEBUG3;
-
-	if( LOG_TO_FILE_AND_CONSOLE::FileReportingLevel() != NONE ) {
-		const std::string pid( std::to_string( getpid() ) );
-
-		fs::path debugFile(DEBUG_DIR);
-		debugFile /= PACKAGE;
-		debugFile += "s-debug-";
-		debugFile += pid;
-		debugFile += ".log";
-
-		errno = 0;
-		_LOGfd = std::fopen(debugFile.string().c_str(), "w");
-
-		if(_LOGfd == nullptr) {
-			LOG(ERROR) << "failed to open debug file";
-			if(errno != 0) {
-				LOG(ERROR) << strerror(errno);
-			}
-		}
-		else {
-			boost::system::error_code ec;
-			fs::permissions(debugFile, fs::owner_read|fs::owner_write|fs::group_read, ec);
-			if( ec.value() != 0 ) {
-				LOG(ERROR) << "failure to set debug file permissions : " << ec.value();
-			}
-		}
-
-		LOG_TO_FILE_AND_CONSOLE::FileStream() = _LOGfd;
-	}
 #endif
-
-	if( _LOGfd == nullptr ) {
-		LOG(INFO) << "debug file not opened";
-	}
 }
 
 DesktopService::~DesktopService() {
@@ -107,12 +75,18 @@ DesktopService::~DesktopService() {
 	LOG(INFO) << GLOGIKS_DESKTOP_SERVICE_NAME << " : bye !";
 	if( _LOGfd != nullptr )
 		std::fclose(_LOGfd);
+
+	closelog();
 }
 
 int DesktopService::run( const int& argc, char *argv[] ) {
-	LOG(INFO) << "Starting " << GLOGIKS_DESKTOP_SERVICE_NAME << " vers. " << VERSION;
-
 	try {
+#if DEBUGGING_ON
+		FileSystem::openDebugFile(GLOGIKS_DESKTOP_SERVICE_NAME, _LOGfd, fs::owner_read|fs::owner_write|fs::group_read);
+#endif
+
+		LOG(INFO) << "Starting " << GLOGIKS_DESKTOP_SERVICE_NAME << " vers. " << VERSION;
+
 		this->parseCommandLine(argc, argv);
 
 		_pid = detachProcess();
@@ -175,10 +149,7 @@ int DesktopService::run( const int& argc, char *argv[] ) {
 	catch ( const GLogiKExcept & e ) {
 		delete _pGKfs; _pGKfs = nullptr;
 
-		std::ostringstream buffer(e.what(), std::ios_base::app);
-		if(errno != 0)
-			buffer << " : " << strerror(errno);
-		LOG(ERROR) << buffer.str();
+		LOG(ERROR) << e.what();
 		return EXIT_FAILURE;
 	}
 	catch ( const GLogiKFatalError & e ) {
@@ -203,7 +174,7 @@ void DesktopService::parseCommandLine(const int& argc, char *argv[]) {
 	try {
 		po::store(po::parse_command_line(argc, argv, desc), vm);
 	}
-	catch( std::exception & e ) {
+	catch( const std::exception & e ) {
 		throw GLogiKExcept( e.what() );
 	}
 	po::notify(vm);