+++ src/lib/dbus/messages/GKDBusErrorReply.cpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2018  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2020  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -107,7 +107,7 @@ void GKDBusMessageErrorReply::sendErrorReply(void) {
 		_errorReply = nullptr;
 	}
 	else {
-		LOG(WARNING) << __func__ << " failure because reply not contructed";
+		LOG(WARNING) << "tried to send NULL error reply";
 	}
 }