+++ src/lib/dbus/messages/GKDBusRemoteMethodCall.cpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2019  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2020  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -165,8 +165,8 @@ void GKDBusMessageRemoteMethodCall::sendRemoteMethodCall(void)
 		_remoteMethodCall = nullptr;
 	}
 	else {
-		LOG(WARNING) << __func__ << " failure because remote_method_call not contructed";
-		throw GKDBusMessageWrongBuild("DBus remote_method_call not contructed");
+		LOG(WARNING) << "tried to send NULL remote method call";
+		throw GKDBusMessageWrongBuild("tried to send NULL remote method call");
 	}
 }
 
@@ -178,7 +178,7 @@ void GKDBusMessageRemoteMethodCall::abandonRemoteMethodCall(void)
 		_remoteMethodCall = nullptr;
 	}
 	else {
-		LOG(WARNING) << __func__ << " failure because remote_method_call not contructed";
+		LOG(WARNING) << "tried to abandon NULL remote method call";
 	}
 }