+++ src/lib/utils/filesystem.cpp
@@ -2,7 +2,7 @@
  *
  *	This file is part of GLogiK project.
  *	GLogiK, daemon to handle special features on gaming keyboards
- *	Copyright (C) 2016-2018  Fabrice Delliaux <netbox253@gmail.com>
+ *	Copyright (C) 2016-2020  Fabrice Delliaux <netbox253@gmail.com>
  *
  *	This program is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
@@ -19,7 +19,15 @@
  *
  */
 
+#include <errno.h>
+
+#include <syslog.h>
+
+#include <cstring>
+#include <cstdlib>
+
 #include <sstream>
+#include <chrono>
 
 #include <config.h>
 
@@ -30,6 +38,8 @@
 #include "filesystem.hpp"
 
 #undef UTILS_COMPILATION
+
+namespace chr = std::chrono;
 
 namespace NSGKUtils
 {
@@ -67,10 +77,20 @@ const std::string FileSystem::getNextAvailableFileName(
 
 		if( mustExist ) {
 			fs::path fullPath = directory / file;
-			if( ! fs::is_regular_file(fullPath) ) {
+			try {
+				if( ! fs::is_regular_file(fullPath) ) {
 #if DEBUGGING_ON
-				LOG(DEBUG3) << "does not exist : " << fullPath.string();
+					LOG(DEBUG3) << "does not exist : " << fullPath.string();
 #endif
+					continue;
+				}
+			}
+			catch (const fs::filesystem_error & e) {
+				LOG(WARNING) << "boost::filesystem::is_regular_file() error : " << e.what();
+				continue;
+			}
+			catch (const std::exception & e) {
+				LOG(WARNING) << "boost::filesystem::is_regular_file() (allocation) error : " << e.what();
 				continue;
 			}
 		}
@@ -84,11 +104,23 @@ const std::string FileSystem::getNextAvailableFileName(
 	throw GLogiKExcept("can't get new filename, counter reached max");
 }
 
-void FileSystem::createOwnerDirectory(const fs::path & directory) {
+void FileSystem::createDirectory(
+	const fs::path & directory,
+	const fs::perms prms)
+{
 #if DEBUGGING_ON
 	bool success = false;
 	LOG(DEBUG2) << "creating directory : " << directory.string();
 #endif
+
+	auto throwError = [] (
+		const std::string & error,
+		const char* what
+			) -> void {
+		std::ostringstream buffer(error, std::ios_base::app);
+		buffer << " : " << what;
+		throw GLogiKExcept( buffer.str() );
+	};
 
 	try {
 #if DEBUGGING_ON
@@ -96,13 +128,16 @@ void FileSystem::createOwnerDirectory(const fs::path & directory) {
 #else
 		fs::create_directory( directory );
 #endif
-		fs::permissions(directory, fs::owner_all);
+
+		if( prms != fs::no_perms ) {
+			fs::permissions(directory, prms);
+		}
 	}
 	catch (const fs::filesystem_error & e) {
-		std::ostringstream buffer;
-		buffer.str("directory creation or set permissions failure : ");
-		buffer << e.what();
-		throw GLogiKExcept( buffer.str() );
+		throwError("directory creation or set permissions failure", e.what());
+	}
+	catch (const std::exception & e) {
+		throwError("directory creation or set permissions (allocation) failure", e.what());
 	}
 
 #if DEBUGGING_ON
@@ -114,6 +149,87 @@ void FileSystem::createOwnerDirectory(const fs::path & directory) {
 	}
 #endif
 }
+
+#if DEBUGGING_ON
+void FileSystem::openDebugFile(
+	const std::string & baseName,
+	FILE* & pFile,
+	const fs::perms prms,
+	const bool createDir
+	)
+{
+	if( LOG_TO_FILE_AND_CONSOLE::FileReportingLevel() == NONE ) {
+		if( pFile == nullptr) {
+			syslog(LOG_INFO, "debug file not opened");
+			return;
+		}
+	}
+
+	fs::path debugFile(DEBUG_DIR);
+
+	if( createDir )
+		FileSystem::createDirectory(debugFile, fs::owner_all | fs::group_all);
+
+	{
+		chr::seconds s = chr::duration_cast< chr::seconds >(
+			chr::system_clock::now().time_since_epoch()
+		);
+
+		debugFile /= baseName;
+		debugFile += "-";
+		debugFile += std::to_string(s.count());
+	}
+
+	debugFile += "-XXXXXX.log";
+
+	int fd = -1;
+	std::string filePath;
+
+	{
+		const std::string f( debugFile.string() );
+		char char_array[f.size() + 1];
+
+		std::strcpy(char_array, f.c_str());
+
+		fd = mkstemps(char_array, 4);
+		filePath = char_array;
+	}
+
+	auto openError = [&filePath] (const std::string & error) -> void {
+		std::ostringstream buffer(std::ios_base::app);
+		buffer << error << " : " << filePath;
+		if(errno != 0) {
+			buffer << " : " << strerror(errno);
+		}
+
+		syslog(LOG_ERR, "%s", buffer.str().c_str());
+	};
+
+	if( fd != -1 ) {
+		pFile = fdopen(fd, "w");
+		if(pFile != nullptr) {
+			if( prms != fs::no_perms ) {
+				boost::system::error_code ec;
+				fs::permissions(filePath, prms, ec);
+				if( ec.value() != 0 ) {
+					std::ostringstream buffer(std::ios_base::app);
+					buffer << "failed to set file permissions : " << filePath
+							<< " : " << ec.message();
+					GKSysLog(LOG_ERR, ERROR, buffer.str());
+				}
+			}
+
+			LOG_TO_FILE_AND_CONSOLE::FileStream() = pFile;
+		}
+		else openError("fdopen error");
+	}
+	else openError("mkstemps error");
+
+	// tried to open debug file but something goes wrong
+	if( pFile == nullptr)
+		syslog(LOG_INFO, "debug file not opened");
+}
+#endif
 
 } // namespace NSGKUtils